+++ SprotGearMall/src/main/java/edu/spring/mall/persistence/LikesDAOImple.java
@@ -4,9 +4,11 @@ import org.apache.ibatis.session.SqlSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
 
 import edu.spring.mall.domain.LikesVO;
 
+@Repository
 public class LikesDAOImple implements LikesDAO {
 	
 	private final Logger logger = LoggerFactory.getLogger(LikesDAOImple.class);
