+++ SprotGearMall/src/main/resources/log4j.xml
@@ -12,7 +12,7 @@
 	
 	<!-- Application Loggers -->
 	<logger name="edu.spring.mall">
-		<level value="debug" />
+		<level value="info" />
 	</logger>
 	
 	<!-- 3rdparty Loggers -->
