+++ SprotGearMall/src/main/webapp/WEB-INF/views/product/detail.jsp
@@ -1,13 +1,16 @@
 <%@page import="edu.spring.mall.domain.ProductVO"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<%@ taglib prefix="sec" uri="http://www.springframework.org/security/tags"%>
+
 <!DOCTYPE html>
 <html>
 <head>
 <script src="https://code.jquery.com/jquery-3.7.1.js"></script>
 <meta charset="UTF-8">
 <title>${vo.productName }</title>
-<link href="<c:url value="/resources/css/headers.css" />" rel="stylesheet">
+<link href="<c:url value="/resources/css/like.css" />" rel="stylesheet">
+
 </head>
 <body>
 	<h2>상품 정보</h2>
@@ -41,7 +44,9 @@
 	<a href="list?page=${page }"><input type="button" value="상품 목록"></a>
 	<a href="update?productName=${vo.productName }&page=${page }"><input type="button" value="상품 수정"></a>
 	<form action="delete" method="POST">
-		<input type="hidden" id="productName" name="productName" value="${vo.productName }">
+		<input type="hidden" id="productId" name="productId" value="${vo.productId }">
+		<input type="hidden" id="memberId" name="memberId" value="${pageContext.request.userPrincipal.name}">
+
 		<input type="submit" value="상품 삭제">
 	</form>
 	
@@ -51,10 +56,47 @@
 	<script type="text/javascript">
 	$(()=>{
 		$('.like-btn').on('click',()=>{
-			$('.heart').toggleClass('heart-filled');
+			$isLike();
 		})
 	})
-	
+	function isLike() {
+    var productId = $('#productId').val();
+    var memberId = $('#memberId').val();
+
+    if ($('.heart').hasClass('heart-filled')) {
+        $.ajax({
+            type: "DELETE", 
+            url: '/likes',
+            headers: { 'Content-Type': 'application/json' },
+            data: JSON.stringify({
+                "memberId": memberId,
+                "productId": productId
+            }),
+            success: (result) => {
+                console.log(result);
+                if (result === 'success') {
+                    $('.heart').toggleClass('heart-filled');
+                }
+            }
+        });
+    } else {
+        $.ajax({
+            type: "POST",
+            url: '/likes',
+            headers: { 'Content-Type': 'application/json' },
+            data: JSON.stringify({
+                "memberId": memberId,
+                "productId": productId
+            }),
+            success: (result) => {
+                console.log(result);
+                if (result === 'success') {
+                    $('.heart').toggleClass('heart-filled');
+                }
+            }
+        });
+    }
+}
 	</script>
 </body>
